+++ cmd/gitserver/internal/git/gitcli/exec_test.go
@@ -1,4 +1,4 @@
-package gitdomain
+package gitcli
 
 import (
 	"fmt"
@@ -81,5 +81,35 @@ func TestIsAllowedDiffGitCmd(t *testing.T) {
 		t.Run(fmt.Sprintf("%s returns %t", strings.Join(cmd.args, " "), cmd.pass), func(t *testing.T) {
 			assert.Equal(t, cmd.pass, IsAllowedGitCmd(logger, cmd.args, "/foo/baz"))
 		})
+	}
+}
+
+func TestStdErrIndicatesCorruption(t *testing.T) {
+	bad := []string{
+		"error: packfile .git/objects/pack/pack-a.pack does not match index",
+		"error: Could not read d24d09b8bc5d1ea2c3aa24455f4578db6aa3afda\n",
+		`error: short SHA1 1325 is ambiguous
+error: Could not read d24d09b8bc5d1ea2c3aa24455f4578db6aa3afda`,
+		`unrelated
+error: Could not read d24d09b8bc5d1ea2c3aa24455f4578db6aa3afda`,
+		"\n\nerror: Could not read d24d09b8bc5d1ea2c3aa24455f4578db6aa3afda",
+		"fatal: commit-graph requires overflow generation data but has none\n",
+		"\rResolving deltas: 100% (21750/21750), completed with 565 local objects.\nfatal: commit-graph requires overflow generation data but has none\nerror: https://github.com/sgtest/megarepo did not send all necessary objects\n\n\": exit status 1",
+	}
+	good := []string{
+		"",
+		"error: short SHA1 1325 is ambiguous",
+		"error: object 156639577dd2ea91cdd53b25352648387d985743 is a blob, not a commit",
+		"error: object 45043b3ff0440f4d7937f8c68f8fb2881759edef is a tree, not a commit",
+	}
+	for _, stderr := range bad {
+		if !stdErrIndicatesCorruption(stderr) {
+			t.Errorf("should contain corrupt line:\n%s", stderr)
+		}
+	}
+	for _, stderr := range good {
+		if stdErrIndicatesCorruption(stderr) {
+			t.Errorf("should not contain corrupt line:\n%s", stderr)
+		}
 	}
 }