+++ cmd/gitserver/internal/server.go
@@ -31,6 +31,7 @@ import (
 	"github.com/sourcegraph/sourcegraph/cmd/gitserver/internal/common"
 	"github.com/sourcegraph/sourcegraph/cmd/gitserver/internal/executil"
 	"github.com/sourcegraph/sourcegraph/cmd/gitserver/internal/git"
+	"github.com/sourcegraph/sourcegraph/cmd/gitserver/internal/git/gitcli"
 	"github.com/sourcegraph/sourcegraph/cmd/gitserver/internal/gitserverfs"
 	"github.com/sourcegraph/sourcegraph/cmd/gitserver/internal/perforce"
 	"github.com/sourcegraph/sourcegraph/cmd/gitserver/internal/urlredactor"
@@ -113,6 +114,8 @@ func NewCloneQueue(obctx *observation.Context, jobs *list.List) *common.Queue[*c
 	return common.NewQueue[*cloneJob](obctx, "clone-queue", jobs)
 }
 
+type Backender func(common.GitDir, api.RepoName) git.GitBackend
+
 // Server is a gitserver server.
 type Server struct {
 	// Logger should be used for all logging and logger creation.
@@ -124,6 +127,10 @@ type Server struct {
 
 	// ReposDir is the path to the base directory for gitserver storage.
 	ReposDir string
+
+	// GetBackendFunc is a function which returns the git backend for a
+	// repository.
+	GetBackendFunc Backender
 
 	// GetRemoteURLFunc is a function which returns the remote URL for a
 	// repository. This is used when cloning or fetching a repository. In
@@ -641,13 +648,6 @@ func (s *Server) ensureOperations() *operations {
 	return s.operations
 }
 
-var blockedCommandExecutedCounter = promauto.NewCounter(prometheus.CounterOpts{
-	Name: "src_gitserver_exec_blocked_command_received",
-	Help: "Incremented each time a command not in the allowlist for gitserver is executed",
-})
-
-var ErrInvalidCommand = errors.New("invalid command")
-
 type NotFoundError struct {
 	Payload *protocol.NotFoundPayload
 }
@@ -665,33 +665,22 @@ type execStatus struct {
 // TODO(@camdencheek): once gRPC is the only consumer of this, do everything with errors
 // because gRPC can handle trailing errors on a stream.
 func (s *Server) exec(ctx context.Context, logger log.Logger, req *protocol.ExecRequest, userAgent string, w io.Writer) (execStatus, error) {
-	// 🚨 SECURITY: Ensure that only commands in the allowed list are executed.
-	// See https://github.com/sourcegraph/security-issues/issues/213.
+	repoName := protocol.NormalizeRepo(req.Repo)
+	dir := gitserverfs.RepoDirFromName(s.ReposDir, repoName)
+	backend := s.GetBackendFunc(dir, repoName)
 
-	repoPath := string(protocol.NormalizeRepo(req.Repo))
-	repoDir := filepath.Join(s.ReposDir, filepath.FromSlash(repoPath))
-
-	if !gitdomain.IsAllowedGitCmd(logger, req.Args, repoDir) {
-		blockedCommandExecutedCounter.Inc()
-		return execStatus{}, ErrInvalidCommand
-	}
-
-	if !req.NoTimeout {
+	if req.NoTimeout {
 		var cancel context.CancelFunc
-		ctx, cancel = context.WithTimeout(ctx, executil.ShortGitCommandTimeout(req.Args))
+		ctx, cancel = context.WithTimeout(ctx, 24*time.Hour)
 		defer cancel()
 	}
 
 	start := time.Now()
 	var cmdStart time.Time // set once we have ensured commit
-	exitStatus := executil.UnsetExitStatus
-	var stdoutN, stderrN int64
 	var status string
 	var execErr error
+	var exitStatus int = executil.UnsetExitStatus
 	ensureRevisionStatus := "noop"
-
-	req.Repo = protocol.NormalizeRepo(req.Repo)
-	repoName := req.Repo
 
 	// Instrumentation
 	{
@@ -714,8 +703,6 @@ func (s *Server) exec(ctx context.Context, logger log.Logger, req *protocol.Exec
 			tr.AddEvent(
 				"done",
 				attribute.String("status", status),
-				attribute.Int64("stdout", stdoutN),
-				attribute.Int64("stderr", stderrN),
 				attribute.String("ensure_revision_status", ensureRevisionStatus),
 			)
 			tr.SetError(execErr)
@@ -746,9 +733,6 @@ func (s *Server) exec(ctx context.Context, logger log.Logger, req *protocol.Exec
 				ev.AddField("ensure_revision_status", ensureRevisionStatus)
 				ev.AddField("client", userAgent)
 				ev.AddField("duration_ms", duration.Milliseconds())
-				ev.AddField("stdin_size", len(req.Stdin))
-				ev.AddField("stdout_size", stdoutN)
-				ev.AddField("stderr_size", stderrN)
 				ev.AddField("exit_status", exitStatus)
 				ev.AddField("status", status)
 				if execErr != nil {
@@ -791,7 +775,6 @@ func (s *Server) exec(ctx context.Context, logger log.Logger, req *protocol.Exec
 		return execStatus{}, &NotFoundError{notFoundPayload}
 	}
 
-	dir := gitserverfs.RepoDirFromName(s.ReposDir, repoName)
 	if s.ensureRevision(ctx, repoName, req.EnsureRevision, dir) {
 		ensureRevisionStatus = "fetched"
 	}
@@ -816,31 +799,33 @@ func (s *Server) exec(ctx context.Context, logger log.Logger, req *protocol.Exec
 		}
 	}
 
-	var stderrBuf bytes.Buffer
-	stdoutW := &writeCounter{w: w}
-	stderrW := &writeCounter{w: &limitWriter{W: &stderrBuf, N: 1024}}
-
 	cmdStart = time.Now()
-	cmd := s.RecordingCommandFactory.Command(ctx, s.Logger, string(repoName), "git", req.Args...)
-	dir.Set(cmd.Unwrap())
-	cmd.Unwrap().Stdout = stdoutW
-	cmd.Unwrap().Stderr = stderrW
-	cmd.Unwrap().Stdin = bytes.NewReader(req.Stdin)
+	stdout, err := backend.Exec(ctx, req.Args...)
+	if err != nil {
+		return execStatus{}, err
+	}
+	defer stdout.Close()
 
-	exitStatus, execErr = executil.RunCommand(ctx, cmd)
+	_, execErr = io.Copy(w, stdout)
+	if execErr != nil {
+		s.logIfCorrupt(ctx, repoName, execErr)
+		commandFailedErr := &gitcli.CommandFailedError{}
+		if errors.As(execErr, &commandFailedErr) {
+			exitStatus = commandFailedErr.ExitStatus
+			status = strconv.Itoa(exitStatus)
+			return execStatus{
+				ExitStatus: commandFailedErr.ExitStatus,
+				Stderr:     string(commandFailedErr.Stderr),
+				Err:        commandFailedErr.Unwrap(),
+			}, nil
+		}
+		return execStatus{}, execErr
+	}
 
+	exitStatus = 0
 	status = strconv.Itoa(exitStatus)
-	stdoutN = stdoutW.n
-	stderrN = stderrW.n
 
-	stderr := stderrBuf.String()
-	s.logIfCorrupt(ctx, repoName, dir, stderr)
-
-	return execStatus{
-		Err:        execErr,
-		Stderr:     stderr,
-		ExitStatus: exitStatus,
-	}, nil
+	return execStatus{ExitStatus: exitStatus}, nil
 }
 
 func setLastFetched(ctx context.Context, db database.DB, shardID string, dir common.GitDir, name api.RepoName) error {
@@ -873,54 +858,13 @@ func (s *Server) setLastErrorNonFatal(ctx context.Context, name api.RepoName, er
 	}
 }
 
-func (s *Server) logIfCorrupt(ctx context.Context, repo api.RepoName, dir common.GitDir, stderr string) {
-	if checkMaybeCorruptRepo(s.Logger, s.RecordingCommandFactory, repo, s.ReposDir, dir, stderr) {
-		reason := stderr
-		if err := s.DB.GitserverRepos().LogCorruption(ctx, repo, reason, s.Hostname); err != nil {
+func (s *Server) logIfCorrupt(ctx context.Context, repo api.RepoName, err error) {
+	var corruptErr common.ErrRepoCorrupted
+	if errors.As(err, &corruptErr) {
+		if err := s.DB.GitserverRepos().LogCorruption(ctx, repo, corruptErr.Reason, s.Hostname); err != nil {
 			s.Logger.Warn("failed to log repo corruption", log.String("repo", string(repo)), log.Error(err))
 		}
 	}
-}
-
-var (
-	// objectOrPackFileCorruptionRegex matches stderr lines from git which indicate
-	// that a repository's packfiles or commit objects might be corrupted.
-	//
-	// See https://github.com/sourcegraph/sourcegraph/issues/6676 for more
-	// context.
-	objectOrPackFileCorruptionRegex = lazyregexp.NewPOSIX(`^error: (Could not read|packfile) `)
-
-	// objectOrPackFileCorruptionRegex matches stderr lines from git which indicate that
-	// git's supplemental commit-graph might be corrupted.
-	//
-	// See https://github.com/sourcegraph/sourcegraph/issues/37872 for more
-	// context.
-	commitGraphCorruptionRegex = lazyregexp.NewPOSIX(`^fatal: commit-graph requires overflow generation data but has none`)
-)
-
-func checkMaybeCorruptRepo(logger log.Logger, rcf *wrexec.RecordingCommandFactory, repo api.RepoName, reposDir string, dir common.GitDir, stderr string) bool {
-	if !stdErrIndicatesCorruption(stderr) {
-		return false
-	}
-
-	logger = logger.With(log.String("repo", string(repo)), log.String("dir", string(dir)))
-	logger.Warn("marking repo for re-cloning due to stderr output indicating repo corruption",
-		log.String("stderr", stderr))
-
-	// We set a flag in the config for the cleanup janitor job to fix. The janitor
-	// runs every minute.
-	err := git.ConfigSet(rcf, reposDir, dir, gitConfigMaybeCorrupt, strconv.FormatInt(time.Now().Unix(), 10))
-	if err != nil {
-		logger.Error("failed to set maybeCorruptRepo config", log.Error(err))
-	}
-
-	return true
-}
-
-// stdErrIndicatesCorruption returns true if the provided stderr output from a git command indicates
-// that there might be repository corruption.
-func stdErrIndicatesCorruption(stderr string) bool {
-	return objectOrPackFileCorruptionRegex.MatchString(stderr) || commitGraphCorruptionRegex.MatchString(stderr)
 }
 
 // testRepoCorrupter is used by tests to disrupt a cloned repository (e.g. deleting
@@ -1128,7 +1072,7 @@ func (s *Server) doClone(
 		testRepoCorrupter(ctx, common.GitDir(tmpPath))
 	}
 
-	if err := postRepoFetchActions(ctx, logger, s.DB, s.Hostname, s.RecordingCommandFactory, s.ReposDir, repo, common.GitDir(tmpPath), remoteURL, syncer); err != nil {
+	if err := postRepoFetchActions(ctx, logger, s.DB, s.Hostname, s.RecordingCommandFactory, repo, common.GitDir(tmpPath), remoteURL, syncer); err != nil {
 		return err
 	}
 
@@ -1220,12 +1164,13 @@ func postRepoFetchActions(
 	db database.DB,
 	shardID string,
 	rcf *wrexec.RecordingCommandFactory,
-	reposDir string,
 	repo api.RepoName,
 	dir common.GitDir,
 	remoteURL *vcs.URL,
 	syncer vcssyncer.VCSSyncer,
 ) (errs error) {
+	backend := gitcli.NewBackend(logger, rcf, dir, repo)
+
 	// Note: We use a multi error in this function to try to make as many of the
 	// post repo fetch actions succeed.
 
@@ -1239,7 +1184,7 @@ func postRepoFetchActions(
 		errs = errors.Append(errs, errors.Wrapf(err, "failed to remove bad refs for repo %q", repo))
 	}
 
-	if err := git.SetRepositoryType(rcf, reposDir, dir, syncer.Type()); err != nil {
+	if err := git.SetRepositoryType(ctx, backend.Config(), syncer.Type()); err != nil {
 		errs = errors.Append(errs, errors.Wrapf(err, "failed to set repository type for repo %q", repo))
 	}
 
@@ -1247,7 +1192,7 @@ func postRepoFetchActions(
 		errs = errors.Append(errs, errors.Wrap(err, "setting git attributes"))
 	}
 
-	if err := gitSetAutoGC(rcf, reposDir, dir); err != nil {
+	if err := gitSetAutoGC(ctx, backend.Config()); err != nil {
 		errs = errors.Append(errs, errors.Wrap(err, "setting git gc mode"))
 	}
 
@@ -1444,9 +1389,9 @@ func (s *Server) doRepoUpdate(ctx context.Context, repo api.RepoName, revspec st
 				}
 
 				// The repo update might have failed due to the repo being corrupt
-				var gitErr *common.GitCommandError
-				if errors.As(err, &gitErr) {
-					s.logIfCorrupt(ctx, repo, gitserverfs.RepoDirFromName(s.ReposDir, repo), gitErr.Output)
+				var corruptErr common.ErrRepoCorrupted
+				if errors.As(err, &corruptErr) {
+					s.logIfCorrupt(ctx, repo, corruptErr)
 				}
 			}
 			s.setLastErrorNonFatal(s.ctx, repo, err)
@@ -1527,7 +1472,7 @@ func (s *Server) doBackgroundRepoUpdate(repo api.RepoName, revspec string) error
 		}
 	}
 
-	return postRepoFetchActions(ctx, logger, s.DB, s.Hostname, s.RecordingCommandFactory, s.ReposDir, repo, dir, remoteURL, syncer)
+	return postRepoFetchActions(ctx, logger, s.DB, s.Hostname, s.RecordingCommandFactory, repo, dir, remoteURL, syncer)
 }
 
 // setHEAD configures git repo defaults (such as what HEAD is) which are